BUG pyximport had no docstring
Transplanted the one from pyximport.pyximport.
